test(crewai): fold crew-chat fixes into test_crew_chat feature suite, drop ticket-named file
